moving concept analysis prep code to service
