+++ public-api/src/main/java/org/pmiops/workbench/cdr/dao/AchillesResultDistDao.java
@@ -11,8 +11,8 @@ public interface AchillesResultDistDao extends CrudRepository<DbAchillesResultDi
     @Query(value = "select * from achilles_results_dist where analysis_id=?1 and stratum_1=?2",nativeQuery=true)
     List<DbAchillesResultDist> fetchConceptDistResults(Long analysisId, String conceptId);
 
-    @Query(value = "select * from achilles_results_dist where analysis_id in (?1) and stratum_1 in (?2)", nativeQuery = true)
-    List<DbAchillesResultDist> fetchByAnalysisIdsAndConceptIds(List<Long> analysisIds,List<String> conceptIds);
+    @Query(value = "select * from achilles_results_dist where analysis_id=?1 and stratum_1 in (?2)", nativeQuery = true)
+    List<DbAchillesResultDist> fetchByAnalysisIdsAndConceptIds(Long analysisId,List<String> conceptIds);
 
 }
