+++ public-api/src/main/java/org/pmiops/workbench/publicapi/DataBrowserController.java
@@ -104,7 +104,7 @@ public class DataBrowserController implements DataBrowserApiDelegate {
     public static final long PARTICIPANT_COUNT_ANALYSIS_ID = 1;
     public static final long COUNT_ANALYSIS_ID = 3000;
     public static final long SURVEY_GENDER_COUNT_ANALYSIS_ID = 3200;
-    public static final long SURVEY_AGE_COUNT_ANALYSIIS_ID = 3201;
+    public static final long SURVEY_AGE_COUNT_ANALYSIS_ID = 3201;
     public static final long GENDER_ANALYSIS_ID = 3101;
     public static final long PARTICIPANT_COUNT_BY_DATE_ANALYSIS_ID = 3107;
     public static final long AGE_ANALYSIS_ID = 3102;
@@ -128,26 +128,15 @@ public class DataBrowserController implements DataBrowserApiDelegate {
 
     public static final long MEASUREMENT_DIST_ANALYSIS_ID = 1815;
 
-    public static final long MEASUREMENT_GENDER_DIST_ANALYSIS_ID = 1815;
-
     public static final long MEASUREMENT_GENDER_ANALYSIS_ID = 1900;
     public static final long MEASUREMENT_GENDER_UNIT_ANALYSIS_ID = 1910;
 
     public static final ArrayList<Long> FMH_CONDITION_CONCEPT_IDS = new ArrayList<>(Arrays.asList(43528515L, 1384639L, 43528634L, 43528761L, 43529158L, 43529767L, 43529272L, 43529217L, 702786L, 43529966L, 43529638L));
     public static final ArrayList<Long> FMH_FM_CONCEPT_IDS = new ArrayList<>(Arrays.asList(43528764L, 43528763L, 43528649L, 43528651L, 43528650L, 43528765L));
 
-    public static final long MALE = 8507;
-    public static final long FEMALE = 8532;
-    public static final long INTERSEX = 1585848;
-    public static final long NONE = 1585849;
-    public static final long OTHER = 0;
-
-    public static final long GENDER_ANALYSIS = 2;
     public static final long RACE_ANALYSIS = 4;
     public static final long ETHNICITY_ANALYSIS = 5;
 
-    public static Map<String, String> ageStratumNameMap = new HashMap<String, String>();
-    public static Map<String, String> genderStratumNameMap = new HashMap<String, String>();
     public static Set<String> validAgeDeciles = new TreeSet<String>(Arrays.asList(new String[]{"2", "3", "4", "5", "6", "7", "8", "9"}));
 
     public DataBrowserController() {}
@@ -169,34 +158,6 @@ public class DataBrowserController implements DataBrowserApiDelegate {
         this.domainInfoService = domainInfoService;
         this.achillesResultService = achillesResultService;
         this.achillesAnalysisService = achillesAnalysisService;
-    }
-
-    public static void setAgeStratumNameMap() {
-        ageStratumNameMap.put("2", "18-29");
-        ageStratumNameMap.put("3", "30-39");
-        ageStratumNameMap.put("4", "40-49");
-        ageStratumNameMap.put("5", "50-59");
-        ageStratumNameMap.put("6", "60-69");
-        ageStratumNameMap.put("7", "70-79");
-        ageStratumNameMap.put("8", "80-89");
-        ageStratumNameMap.put("9", "89+");
-    }
-
-    public static void setGenderStratumNameMap() {
-        /* This is to slow to use the db */
-        genderStratumNameMap.put("8507", "Male");
-        genderStratumNameMap.put("8532", "Female");
-        genderStratumNameMap.put("8521", "Other");
-        genderStratumNameMap.put("8551", "Unknown");
-        genderStratumNameMap.put("8570", "Ambiguous");
-        genderStratumNameMap.put("1585849", "None of these describe me");
-        genderStratumNameMap.put("1585848", "Intersex");
-        genderStratumNameMap.put("0", "Other");
-    }
-
-    static {
-        setAgeStratumNameMap();
-        setGenderStratumNameMap();
     }
 
     /**
@@ -276,28 +237,6 @@ public class DataBrowserController implements DataBrowserApiDelegate {
                             .conceptId(criteria.getConceptId())
                             .path(criteria.getPath())
                             .canSelect(criteria.getCanSelect());
-                }
-            };
-
-    /**
-     * Converter function from backend representation (used with Hibernate) to
-     * client representation (generated by Swagger).
-     */
-    private static final Function<ConceptAnalysis, ConceptAnalysis>
-            TO_CLIENT_CONCEPTANALYSIS=
-            new Function<ConceptAnalysis, ConceptAnalysis>() {
-                @Override
-                public ConceptAnalysis apply(ConceptAnalysis ca) {
-                    return new ConceptAnalysis()
-                            .conceptId(ca.getConceptId())
-                            .countAnalysis(ca.getCountAnalysis())
-                            .genderAnalysis(ca.getGenderAnalysis())
-                            .ageAnalysis(ca.getAgeAnalysis())
-                            .measurementValueGenderAnalysis(ca.getMeasurementValueGenderAnalysis())
-                            .measurementValueAgeAnalysis(ca.getMeasurementValueAgeAnalysis())
-                            .measurementDistributionAnalysis(ca.getMeasurementDistributionAnalysis())
-                            .measurementGenderCountAnalysis(ca.getMeasurementGenderCountAnalysis())
-                            .participantCountAnalysis(ca.getParticipantCountAnalysis());
                 }
             };
 
@@ -549,10 +488,7 @@ public class DataBrowserController implements DataBrowserApiDelegate {
         } catch(NullPointerException ie) {
             throw new ServerErrorException("Cannot set default cdr version");
         }
-        Analysis genderAnalysis = achillesAnalysisService.findAnalysisById(GENDER_ANALYSIS);
-        Optional.ofNullable(genderAnalysis).orElseThrow(() -> new DataNotFoundException("Gender Analysis data is not available"));
-        addGenderStratum(genderAnalysis,1, "0", null);
-        return ResponseEntity.ok(genderAnalysis);
+        return ResponseEntity.ok(achillesAnalysisService.getGenderAnalysis());
     }
 
     @Override
@@ -705,41 +641,16 @@ public class DataBrowserController implements DataBrowserApiDelegate {
     @Override
     public ResponseEntity<CountAnalysis> getCountAnalysis(String domainId, String domainDesc) {
         cdrVersionService.setDefaultCdrVersion();
-
-        if (domainDesc.equals("ehr") || domainDesc.equals("fitbit")) {
-            List<Analysis> ehrAnalysesList = achillesAnalysisService.findAnalysisByIdsAndDomain(ImmutableList.of(3300L, 3301L), domainId);
-            CountAnalysis ehrCountAnalysis = new CountAnalysis();
-            ehrCountAnalysis.setDomainId(domainId);
-            Analysis genderCountAnalysis = ehrAnalysesList.stream().filter(aa -> aa.getAnalysisId() == 3300).collect(Collectors.toList()).get(0);
-            Analysis ageCountAnalysis = ehrAnalysesList.stream().filter(aa -> aa.getAnalysisId() == 3301).collect(Collectors.toList()).get(0);
-            addGenderStratum(genderCountAnalysis,4, domainId, null);
-            addAgeStratum(ageCountAnalysis, domainId, null,  4);
-            ehrCountAnalysis.setGenderCountAnalysis(genderCountAnalysis);
-            ehrCountAnalysis.setAgeCountAnalysis(ageCountAnalysis);
-            return ResponseEntity.ok(ehrCountAnalysis);
-        } else if (domainDesc.equals("survey")){
-            List<Analysis> ehrAnalysesList = achillesAnalysisService.findSurveyAnalysisByIds(ImmutableList.of(SURVEY_GENDER_COUNT_ANALYSIS_ID, SURVEY_AGE_COUNT_ANALYSIIS_ID), domainId);
-            CountAnalysis surveyCountAnalysis = new CountAnalysis();
-            surveyCountAnalysis.setDomainId(domainId);
-            Analysis genderCountAnalysis = ehrAnalysesList.stream().filter(aa -> aa.getAnalysisId() == SURVEY_GENDER_COUNT_ANALYSIS_ID).collect(Collectors.toList()).get(0);
-            Analysis ageCountAnalysis = ehrAnalysesList.stream().filter(aa -> aa.getAnalysisId() == SURVEY_AGE_COUNT_ANALYSIIS_ID).collect(Collectors.toList()).get(0);
-            addGenderStratum(genderCountAnalysis,2, domainId, null);
-            addAgeStratum(ageCountAnalysis, domainId, null,  2);
-            surveyCountAnalysis.setGenderCountAnalysis(genderCountAnalysis);
-            surveyCountAnalysis.setAgeCountAnalysis(ageCountAnalysis);
-            return ResponseEntity.ok(surveyCountAnalysis);
+        ImmutableList<Long> analysisIds = null;
+        int stratum;
+        if (domainDesc.equals("survey")){
+            analysisIds = ImmutableList.of(SURVEY_GENDER_COUNT_ANALYSIS_ID, SURVEY_AGE_COUNT_ANALYSIS_ID);
+            stratum = 2;
         } else {
-            List<Analysis> ehrAnalysesList = achillesAnalysisService.findAnalysisByIdsAndDomain(ImmutableList.of(3300L, 3301L), domainId);
-            CountAnalysis ehrCountAnalysis = new CountAnalysis();
-            ehrCountAnalysis.setDomainId(domainId);
-            Analysis genderCountAnalysis = ehrAnalysesList.stream().filter(aa -> aa.getAnalysisId() == 3300).collect(Collectors.toList()).get(0);
-            Analysis ageCountAnalysis = ehrAnalysesList.stream().filter(aa -> aa.getAnalysisId() == 3301).collect(Collectors.toList()).get(0);
-            addGenderStratum(genderCountAnalysis,4, domainId, null);
-            addAgeStratum(ageCountAnalysis, domainId, null,  4);
-            ehrCountAnalysis.setGenderCountAnalysis(genderCountAnalysis);
-            ehrCountAnalysis.setAgeCountAnalysis(ageCountAnalysis);
-            return ResponseEntity.ok(ehrCountAnalysis);
+            analysisIds = ImmutableList.of(EHR_GENDER_COUNT_ANALYSIS_ID, EHR_AGE_COUNT_ANALYSIS_ID);
+            stratum = 4;
         }
+        return ResponseEntity.ok(achillesAnalysisService.getCountAnalysis(domainId, domainDesc, analysisIds, stratum));
     }
 
     @Override
@@ -765,142 +676,7 @@ public class DataBrowserController implements DataBrowserApiDelegate {
         }
 
         ConceptAnalysisListResponse resp=new ConceptAnalysisListResponse();
-        List<ConceptAnalysis> conceptAnalysisList=new ArrayList<>();
-
-        List<DbAchillesResultDist> overallDistResults = achillesResultDistService.fetchByAnalysisIdsAndConceptIds(new ArrayList<Long>( Arrays.asList(MEASUREMENT_GENDER_DIST_ANALYSIS_ID) ),conceptIds);
-
-        Multimap<Long, DbAchillesResultDist> distResultsByAnalysisId = null;
-        if(overallDistResults != null){
-            distResultsByAnalysisId = Multimaps
-                    .index(overallDistResults, DbAchillesResultDist::getAnalysisId);
-        }
-
-        HashMap<Long,HashMap<String,List<DbAchillesResultDist>>> dbAnalysisDistResults = new HashMap<>();
-
-        for(Long key:distResultsByAnalysisId.keySet()){
-            Multimap<String, DbAchillesResultDist> conceptDistResults = Multimaps.index(distResultsByAnalysisId.get(key), DbAchillesResultDist::getStratum1);
-            for(String concept:conceptDistResults.keySet()) {
-                if(dbAnalysisDistResults.containsKey(key)){
-                    HashMap<String,List<DbAchillesResultDist>> results = dbAnalysisDistResults.get(key);
-                    results.put(concept,new ArrayList<>(conceptDistResults.get(concept)));
-                }else{
-                    HashMap<String,List<DbAchillesResultDist>> results = new HashMap<>();
-                    results.put(concept,new ArrayList<>(conceptDistResults.get(concept)));
-                    dbAnalysisDistResults.put(key,results);
-                }
-            }
-        }
-
-        for(String conceptId: conceptIds){
-            ConceptAnalysis conceptAnalysis=new ConceptAnalysis();
-
-            boolean isMeasurement = false;
-
-                List<Analysis> analysisList = achillesAnalysisService.findConceptAnalysisResults(conceptId,ImmutableList.of(GENDER_ANALYSIS_ID, AGE_ANALYSIS_ID, COUNT_ANALYSIS_ID, MEASUREMENT_GENDER_ANALYSIS_ID, MEASUREMENT_DIST_ANALYSIS_ID, MEASUREMENT_GENDER_UNIT_ANALYSIS_ID));
-
-                HashMap<Long, Analysis> analysisHashMap = new HashMap<>();
-                for (Analysis aa : analysisList) {
-                    analysisHashMap.put(aa.getAnalysisId(), aa);
-                }
-
-                conceptAnalysis.setConceptId(conceptId);
-                Iterator it = analysisHashMap.entrySet().iterator();
-                while (it.hasNext()) {
-                    Map.Entry pair = (Map.Entry) it.next();
-                    Long analysisId = (Long) pair.getKey();
-                    Analysis aa = (Analysis) pair.getValue();
-                    //aa.setUnitName(unitName);
-                    if (analysisId != MEASUREMENT_GENDER_UNIT_ANALYSIS_ID && analysisId != MEASUREMENT_GENDER_ANALYSIS_ID && analysisId != MEASUREMENT_DIST_ANALYSIS_ID && !Strings.isNullOrEmpty(domainId)) {
-                        aa.setResults(aa.getResults().stream().filter(ar -> ar.getStratum3().equalsIgnoreCase(domainId)).collect(Collectors.toList()));
-                    }
-                    if (analysisId == COUNT_ANALYSIS_ID) {
-                        conceptAnalysis.setCountAnalysis(aa);
-                    } else if (analysisId == GENDER_ANALYSIS_ID) {
-                        addGenderStratum(aa, 2, conceptId, null);
-                        conceptAnalysis.setGenderAnalysis(aa);
-                    } else if (analysisId == AGE_ANALYSIS_ID) {
-                        addAgeStratum(aa, conceptId, null, 2);
-                        conceptAnalysis.setAgeAnalysis(aa);
-                    } else if (analysisId == MEASUREMENT_GENDER_ANALYSIS_ID) {
-                        Map<String, List<AchillesResult>> results = seperateUnitResults(aa);
-                        List<Analysis> unitSeperateAnalysis = new ArrayList<>();
-                        HashMap<String, List<DbAchillesResultDist>> distResults = dbAnalysisDistResults.get(MEASUREMENT_GENDER_DIST_ANALYSIS_ID);
-                        if (distResults != null) {
-                            List<DbAchillesResultDist> conceptDistResults = distResults.get(conceptId);
-                            if (conceptDistResults != null) {
-                                Multimap<String, DbAchillesResultDist> unitDistResults = Multimaps.index(conceptDistResults, DbAchillesResultDist::getStratum2);
-                                for (String unit : unitDistResults.keySet()) {
-                                    if (results.keySet().contains(unit)) {
-                                        Analysis unitGenderAnalysis = achillesMapper.makeCopyAnalysis(aa);
-                                        unitGenderAnalysis.setResults(results.get(unit));
-                                        unitGenderAnalysis.setUnitName(unit);
-                                        if (!unit.equalsIgnoreCase("no unit")) {
-                                            processMeasurementGenderMissingBins(MEASUREMENT_GENDER_DIST_ANALYSIS_ID, unitGenderAnalysis, conceptId, unit, new ArrayList<>(unitDistResults.get(unit)), "numeric");
-                                        } else {
-                                            //Seperate text and numeric values
-                                            ArrayList<AchillesResult> textValues = new ArrayList<>();
-                                            ArrayList<AchillesResult> numericValues = new ArrayList<>();
-                                            // In case no unit has a mix of text and numeric values, only display text values as mix does not make sense to user.
-                                            for (AchillesResult result : unitGenderAnalysis.getResults()) {
-                                                if (result.getStratum5() == null || result.getStratum5().trim().isEmpty()) {
-                                                    result.setMeasurementValueType("numeric");
-                                                    numericValues.add(result);
-                                                } else {
-                                                    result.setMeasurementValueType("text");
-                                                    textValues.add(result);
-                                                }
-                                            }
-
-                                            if (textValues.size() > 0) {
-                                                processMeasurementGenderMissingBins(MEASUREMENT_GENDER_DIST_ANALYSIS_ID, unitGenderAnalysis, conceptId, null, null, "text");
-                                            }
-                                            if (numericValues.size() > 0) {
-                                                processMeasurementGenderMissingBins(MEASUREMENT_GENDER_DIST_ANALYSIS_ID, unitGenderAnalysis, conceptId, null, null, "numeric");
-                                            }
-                                            unitGenderAnalysis.setResults(results.get(unit));
-                                            unitGenderAnalysis.setUnitName(unit);
-
-                                        }
-                                        unitSeperateAnalysis.add(unitGenderAnalysis);
-                                    }
-                                }
-                            } else {
-                                unitSeperateAnalysis.add(aa);
-                            }
-                        }
-                        addGenderStratum(aa, 3, conceptId, null);
-                        isMeasurement = true;
-                        conceptAnalysis.setMeasurementValueGenderAnalysis(unitSeperateAnalysis);
-                    } else if (analysisId == MEASUREMENT_GENDER_UNIT_ANALYSIS_ID) {
-                        Map<String, List<AchillesResult>> results = seperateUnitResults(aa);
-                        List<Analysis> unitSeperateAnalysis = new ArrayList<>();
-                        for (String unit : results.keySet()) {
-                            Analysis unitGenderCountAnalysis = achillesMapper.makeCopyAnalysis(aa);
-                            unitGenderCountAnalysis.setResults(results.get(unit));
-                            unitGenderCountAnalysis.setUnitName(unit);
-                            unitSeperateAnalysis.add(unitGenderCountAnalysis);
-                        }
-                        isMeasurement = true;
-                        conceptAnalysis.setMeasurementGenderCountAnalysis(unitSeperateAnalysis);
-                    }
-                }
-
-                if (isMeasurement) {
-                    Analysis measurementDistAnalysis = achillesAnalysisService.findAnalysisById(MEASUREMENT_DIST_ANALYSIS_ID);
-                    List<AchillesResultDist> achillesResultDistList = achillesResultDistService.fetchConceptDistResults(MEASUREMENT_DIST_ANALYSIS_ID, conceptId);
-                    HashMap<String, List<AchillesResultDist>> results = seperateDistResultsByUnit(achillesResultDistList);
-                    List<Analysis> unitSeperateAnalysis = new ArrayList<>();
-                    for (String unit : results.keySet()) {
-                        Analysis mDistAnalysis = achillesMapper.makeCopyAnalysis(measurementDistAnalysis);
-                        mDistAnalysis.setDistResults(results.get(unit));
-                        mDistAnalysis.setUnitName(unit);
-                        unitSeperateAnalysis.add(mDistAnalysis);
-                    }
-                    conceptAnalysis.setMeasurementDistributionAnalysis(unitSeperateAnalysis);
-                }
-                conceptAnalysisList.add(conceptAnalysis);
-        }
-        resp.setItems(conceptAnalysisList.stream().map(TO_CLIENT_CONCEPTANALYSIS).collect(Collectors.toList()));
+        resp.setItems(achillesAnalysisService.getConceptAnalyses(conceptIds, domainId));
         return ResponseEntity.ok(resp);
     }
 
@@ -913,46 +689,7 @@ public class DataBrowserController implements DataBrowserApiDelegate {
         }
 
         ConceptAnalysisListResponse resp=new ConceptAnalysisListResponse();
-        List<ConceptAnalysis> conceptAnalysisList=new ArrayList<>();
-
-        List<Analysis> analysisList = achillesAnalysisService.findAnalysisByIdsAndDomain(ImmutableList.of(GENDER_ANALYSIS_ID, AGE_ANALYSIS_ID, COUNT_ANALYSIS_ID, PARTICIPANT_COUNT_BY_DATE_ANALYSIS_ID), "Fitbit");
-
-        if (analysisList.size() == 0) {
-            throw new DataNotFoundException("Cannot find analysis data of this concept");
-        }
-
-        HashMap<Long, Analysis> analysisHashMap = new HashMap<>();
-        for(Analysis aa: analysisList){
-            analysisHashMap.put(aa.getAnalysisId(), aa);
-        }
-
-        for (String concept: concepts) {
-
-            ConceptAnalysis conceptAnalysis=new ConceptAnalysis();
-
-            Analysis countAnalysis = achillesMapper.makeCopyAnalysis(analysisHashMap.get(COUNT_ANALYSIS_ID));
-            Analysis ageAnalysis = achillesMapper.makeCopyAnalysis(analysisHashMap.get(AGE_ANALYSIS_ID));
-            Analysis genderAnalysis = achillesMapper.makeCopyAnalysis(analysisHashMap.get(GENDER_ANALYSIS_ID));
-            Analysis participantCountAnalysis = achillesMapper.makeCopyAnalysis(analysisHashMap.get(PARTICIPANT_COUNT_BY_DATE_ANALYSIS_ID));
-
-            countAnalysis.setResults(analysisHashMap.get(COUNT_ANALYSIS_ID).getResults());
-            ageAnalysis.setResults(analysisHashMap.get(AGE_ANALYSIS_ID).getResults().stream().filter(ar -> ar.getStratum1().equals(concept)).collect(Collectors.toList()));
-            genderAnalysis.setResults(analysisHashMap.get(GENDER_ANALYSIS_ID).getResults().stream().filter(ar -> ar.getStratum1().equals(concept)).collect(Collectors.toList()));
-            participantCountAnalysis.setResults(analysisHashMap.get(PARTICIPANT_COUNT_BY_DATE_ANALYSIS_ID).getResults().stream().filter(ar -> ar.getStratum1().equals(concept)).collect(Collectors.toList()));
-            participantCountAnalysis.getResults().sort((o1, o2) -> o1.getStratum2().compareTo(o2.getStratum2()));
-
-            addGenderStratum(genderAnalysis,2, concept, null);
-            addAgeStratum(ageAnalysis, concept, null, 2);
-
-            conceptAnalysis.setConceptId(concept);
-            conceptAnalysis.setCountAnalysis(countAnalysis);
-            conceptAnalysis.setGenderAnalysis(genderAnalysis);
-            conceptAnalysis.setAgeAnalysis(ageAnalysis);
-            conceptAnalysis.setParticipantCountAnalysis(participantCountAnalysis);
-            conceptAnalysisList.add(conceptAnalysis);
-        }
-
-        resp.setItems(conceptAnalysisList.stream().map(TO_CLIENT_CONCEPTANALYSIS).collect(Collectors.toList()));
+        resp.setItems(achillesAnalysisService.getFitbitConceptAnalyses(concepts));
         return ResponseEntity.ok(resp);
     }
 
@@ -987,440 +724,6 @@ public class DataBrowserController implements DataBrowserApiDelegate {
             throw new ServerErrorException("Cannot set default cdr version");
         }
         return ResponseEntity.ok(achillesResultService.findAchillesResultByAnalysisId(PARTICIPANT_COUNT_ANALYSIS_ID));
-    }
-
-    public ArrayList<Float> makeBins(Float min,Float max, Float binWidth) {
-        TreeSet<String> bins = new TreeSet<>();
-
-        bins.add(String.format("%.2f", min));
-
-        if (min+binWidth <= max) {
-            bins.add(String.format("%.2f", min+binWidth));
-        }
-        if (min+2*binWidth <= max) {
-            bins.add(String.format("%.2f", min+2*binWidth));
-        }
-        if (min+3*binWidth <= max) {
-            bins.add(String.format("%.2f", min+3*binWidth));
-        }
-        if (min+4*binWidth <= max) {
-            bins.add(String.format("%.2f", min+4*binWidth));
-        }
-        if (min+5*binWidth <= max) {
-            bins.add(String.format("%.2f", min+5*binWidth));
-        }
-        if (min+6*binWidth <= max) {
-            bins.add(String.format("%.2f", min+6*binWidth));
-        }
-        if (min+7*binWidth <= max) {
-            bins.add(String.format("%.2f", min+7*binWidth));
-        }
-        if (min+8*binWidth <= max) {
-            bins.add(String.format("%.2f", min+8*binWidth));
-        }
-        if (min+9*binWidth <= max) {
-            bins.add(String.format("%.2f", min+9*binWidth));
-        }
-        if (min+10*binWidth <= max) {
-            bins.add(String.format("%.2f", min+10*binWidth));
-        }
-        if (min+11*binWidth <= max) {
-            bins.add(String.format("%.2f", min+11*binWidth));
-        }
-
-        List<String> trimmedBins = new ArrayList<>();
-
-        for(String s:bins) {
-            trimmedBins.add(s.indexOf(".") < 0 ? s : s.replaceAll("0*$", "").replaceAll("\\.$", ""));
-        }
-
-        return new ArrayList<Float>(trimmedBins.stream().map(Float::parseFloat).collect(Collectors.toList()));
-    }
-
-    public void addGenderStratum(Analysis aa, int stratum, String conceptId, List<DbAchillesResult> ehrCountResults){
-        Set<String> uniqueGenderStratums = new TreeSet<String>();
-        String domainConceptId = null;
-        for(AchillesResult ar: aa.getResults()){
-            String analysisStratumName =ar.getAnalysisStratumName();
-            if (analysisStratumName == null || analysisStratumName.equals("")) {
-                if (stratum == 1) {
-                    uniqueGenderStratums.add(ar.getStratum1());
-                    ar.setAnalysisStratumName(genderStratumNameMap.get(ar.getStratum1()));
-                } else if (stratum == 2) {
-                    uniqueGenderStratums.add(ar.getStratum2());
-                    ar.setAnalysisStratumName(genderStratumNameMap.get(ar.getStratum2()));
-                } else if (stratum == 3) {
-                    uniqueGenderStratums.add(ar.getStratum3());
-                    ar.setAnalysisStratumName(genderStratumNameMap.get(ar.getStratum3()));
-                } else if (stratum == 4) {
-                    domainConceptId = ar.getStratum1();
-                    uniqueGenderStratums.add(ar.getStratum4());
-                    ar.setAnalysisStratumName(genderStratumNameMap.get(ar.getStratum4()));
-                }
-            }
-        }
-        if(uniqueGenderStratums.size() < 3) {
-            Set<String> completeGenderStratumList = new TreeSet<String>(Arrays.asList(new String[] {"8507", "8532", "0"}));
-            completeGenderStratumList.removeAll(uniqueGenderStratums);
-            for(String missingGender: completeGenderStratumList){
-                AchillesResult missingResult = null;
-                if (aa.getAnalysisId() == EHR_GENDER_COUNT_ANALYSIS_ID) {
-                    missingResult = achillesMapper.makeCopyAchillesResult(aa.getAnalysisId(), domainConceptId, null, conceptId, missingGender, null, null,null, 20L, 20L);
-                } else {
-                    if (stratum == 1) {
-                        missingResult = achillesMapper.makeCopyAchillesResult(aa.getAnalysisId(), missingGender, null, null, null, null, null,null, 20L, 20L);
-                    } else if (stratum == 2) {
-                        missingResult = achillesMapper.makeCopyAchillesResult(aa.getAnalysisId(), conceptId, missingGender, null, null, null, null,null, 20L, 20L);
-                    } else if (stratum == 3) {
-                        missingResult = achillesMapper.makeCopyAchillesResult(aa.getAnalysisId(), conceptId, null, missingGender, null, null, null,null, 20L, 20L);
-                    }
-                }
-                missingResult.setAnalysisStratumName(genderStratumNameMap.get(missingGender));
-                aa.getResults().add(missingResult);
-            }
-        }
-        aa.setResults(aa.getResults().stream().filter(ar -> ar.getAnalysisStratumName() != null).collect(Collectors.toList()));
-    }
-
-    public void addAgeStratum(Analysis aa, String conceptId, List<DbAchillesResult> ehrCountResults, int stratum){
-        Set<String> uniqueAgeDeciles = new TreeSet<String>();
-        for(AchillesResult ar: aa.getResults()){
-            String analysisStratumName=ar.getAnalysisStratumName();
-            if (analysisStratumName == null || analysisStratumName.equals("")) {
-                if (stratum == 2) {
-                    if (ar.getStratum2() != null && !ar.getStratum2().equals("0")) {
-                        uniqueAgeDeciles.add(ar.getStratum2());
-                        ar.setAnalysisStratumName(ageStratumNameMap.get(ar.getStratum2()));
-                    }
-                } else if (stratum == 4) {
-                    if (ar.getStratum4() != null && !ar.getStratum4().equals("0")) {
-                        uniqueAgeDeciles.add(ar.getStratum4());
-                        ar.setAnalysisStratumName(ageStratumNameMap.get(ar.getStratum4()));
-                    }
-                }
-            }
-        }
-        aa.setResults(aa.getResults().stream().filter(ar -> ar.getAnalysisStratumName() != null).collect(Collectors.toList()));
-        if(uniqueAgeDeciles.size() < 8){
-            if (aa.getAnalysisId() == EHR_AGE_COUNT_ANALYSIS_ID) {
-                Set<String> completeAgeDeciles = new TreeSet<String>(Arrays.asList(new String[] {"2", "3", "4", "5", "6", "7", "8", "9"}));
-                completeAgeDeciles.removeAll(uniqueAgeDeciles);
-                for(String missingAgeDecile: completeAgeDeciles){
-                    List<DbAchillesResult> ehrAgeCountResults = null;
-                    try {
-                        ehrAgeCountResults  = ehrCountResults.stream().filter(ar -> ar.getStratum4().equals(missingAgeDecile)).collect(Collectors.toList());
-                    } catch(NullPointerException npe) {
-                    }
-                    AchillesResult missingResult = null;
-                    if (ehrAgeCountResults != null && ehrAgeCountResults.size() > 0) {
-                        DbAchillesResult result = ehrAgeCountResults.get(0);
-                        String percentageValue = String.valueOf(Math.round((20/result.getCountValue())*100/2)*2);
-                        missingResult = achillesMapper.makeCopyAchillesResult(EHR_AGE_COUNT_ANALYSIS_ID, conceptId, null, null, missingAgeDecile, null, null, null, 20L, 20L);
-                    } else {
-                        missingResult = achillesMapper.makeCopyAchillesResult(EHR_AGE_COUNT_ANALYSIS_ID, conceptId, null, null, missingAgeDecile, null, null, null, 20L, 20L);
-                    }
-                    missingResult.setAnalysisStratumName(ageStratumNameMap.get(missingAgeDecile));
-                    aa.getResults().add(missingResult);
-                }
-            } else {
-                Set<String> completeAgeDeciles = new TreeSet<String>(Arrays.asList(new String[] {"2", "3", "4", "5", "6", "7", "8", "9"}));
-                completeAgeDeciles.removeAll(uniqueAgeDeciles);
-                for(String missingAgeDecile: completeAgeDeciles){
-                    AchillesResult missingResult = achillesMapper.makeCopyAchillesResult(AGE_ANALYSIS_ID, conceptId, missingAgeDecile, null, null, null, null, null, 20L, 20L);
-                    missingResult.setAnalysisStratumName(ageStratumNameMap.get(missingAgeDecile));
-                    aa.getResults().add(missingResult);
-                }
-            }
-        }
-    }
-
-    public void processMeasurementGenderMissingBins(Long analysisId, Analysis aa, String conceptId, String unitName, List<DbAchillesResultDist> resultDists, String type) {
-
-        if (resultDists != null) {
-            Float maleBinMin = null;
-            Float maleBinMax = null;
-
-            Float femaleBinMin = null;
-            Float femaleBinMax = null;
-
-            Float otherBinMin = null;
-            Float otherBinMax = null;
-
-            float maleBinWidth = 0f;
-            float femaleBinWidth = 0f;
-            float otherBinWidth = 0f;
-
-            for(DbAchillesResultDist ard:resultDists){
-                if(Integer.parseInt(ard.getStratum3())== MALE) {
-                    maleBinMin = Float.valueOf(ard.getStratum4());
-                    maleBinMax = Float.valueOf(ard.getStratum5());
-                    maleBinWidth = Float.valueOf(ard.getStratum6());
-
-                }
-                else if(Integer.parseInt(ard.getStratum3()) == FEMALE) {
-                    femaleBinMin = Float.valueOf(ard.getStratum4());
-                    femaleBinMax = Float.valueOf(ard.getStratum5());
-                    femaleBinWidth = Float.valueOf(ard.getStratum6());
-                }
-                else if(Integer.parseInt(ard.getStratum3()) == OTHER) {
-                    otherBinMin = Float.valueOf(ard.getStratum4());
-                    otherBinMax = Float.valueOf(ard.getStratum5());
-                    otherBinWidth = Float.valueOf(ard.getStratum6());
-                }
-            }
-
-            if (femaleBinMax == null && femaleBinMin == null) {
-                if (maleBinMin != null && maleBinMax != null && otherBinMax != null && otherBinMin != null) {
-                    femaleBinMin = Math.min(maleBinMin, otherBinMin);
-                    femaleBinMax = Math.max(maleBinMax, otherBinMax);
-                    femaleBinWidth = normalizeBinWidth(femaleBinMin, femaleBinMax);
-                } else if (maleBinMin != null && maleBinMax != null) {
-                    femaleBinMin = maleBinMin;
-                    femaleBinMax = maleBinMax;
-                    femaleBinWidth = maleBinWidth;
-                } else if (otherBinMax != null && otherBinMin != null) {
-                    femaleBinMin = otherBinMin;
-                    femaleBinMax = otherBinMax;
-                    femaleBinWidth = otherBinWidth;
-                }
-            }
-
-            if (maleBinMax == null && maleBinMin == null) {
-                if (femaleBinMin != null && femaleBinMax != null && otherBinMax != null && otherBinMin != null) {
-                    maleBinMin = Math.min(femaleBinMin, otherBinMin);
-                    maleBinMax = Math.max(femaleBinMax, otherBinMax);
-                    maleBinWidth = normalizeBinWidth(maleBinMin, maleBinMax);
-                } else if (femaleBinMin != null && femaleBinMax != null) {
-                    maleBinMin = femaleBinMin;
-                    maleBinMax = femaleBinMax;
-                    maleBinWidth = femaleBinWidth;
-                } else if (otherBinMax != null && otherBinMin != null) {
-                    maleBinMin = otherBinMin;
-                    maleBinMax = otherBinMax;
-                    maleBinWidth = otherBinWidth;
-                }
-            }
-
-            if (otherBinMax == null && otherBinMin == null) {
-                if (femaleBinMin != null && femaleBinMax != null && maleBinMax != null && maleBinMin != null) {
-                    otherBinMin = Math.min(femaleBinMin, maleBinMin);
-                    otherBinMax = Math.max(femaleBinMax, maleBinMax);
-                    otherBinWidth = normalizeBinWidth(otherBinMin, otherBinMax);
-                } else if (femaleBinMin != null && femaleBinMax != null) {
-                    otherBinMin = femaleBinMin;
-                    otherBinMax = femaleBinMax;
-                    otherBinWidth = femaleBinWidth;
-                } else if (maleBinMin != null && maleBinMax != null) {
-                    otherBinMax = maleBinMax;
-                    otherBinMin = maleBinMin;
-                    otherBinWidth = maleBinWidth;
-                }
-            }
-
-            ArrayList<Float> maleBinRanges = new ArrayList<Float>();
-            ArrayList<Float> femaleBinRanges = new ArrayList<Float>();
-            ArrayList<Float> otherBinRanges = new ArrayList<Float>();
-
-            if(maleBinMax != null && maleBinMin != null){
-                maleBinRanges = makeBins(maleBinMin, maleBinMax, maleBinWidth);
-            }
-
-            if(femaleBinMax != null && femaleBinMin != null){
-                femaleBinRanges = makeBins(femaleBinMin, femaleBinMax, femaleBinWidth);
-            }
-
-            if(otherBinMax != null && otherBinMin != null){
-                otherBinRanges = makeBins(otherBinMin, otherBinMax, otherBinWidth);
-            }
-
-            Collections.sort(maleBinRanges);
-            Collections.sort(femaleBinRanges);
-            Collections.sort(otherBinRanges);
-
-            ArrayList<String> maleRangesInResults = new ArrayList<>();
-            maleRangesInResults.add(("< " + trimTrailingZeroDecimals(String.valueOf(maleBinRanges.get(0)))));
-            for (int k=0; k<maleBinRanges.size()-1;k++) {
-                maleRangesInResults.add((trimTrailingZeroDecimals(String.valueOf(maleBinRanges.get(k))) + " - " + trimTrailingZeroDecimals(String.valueOf(maleBinRanges.get(k+1)))));
-            }
-            maleRangesInResults.add((">= " + trimTrailingZeroDecimals(String.valueOf(maleBinRanges.get(maleBinRanges.size()-1)))));
-
-            ArrayList<String> femaleRangesInResults = new ArrayList<>();
-            femaleRangesInResults.add(("< " + trimTrailingZeroDecimals(String.valueOf(femaleBinRanges.get(0)))));
-            for (int k=0; k<femaleBinRanges.size()-1;k++) {
-                femaleRangesInResults.add((trimTrailingZeroDecimals(String.valueOf(femaleBinRanges.get(k))) + " - " + trimTrailingZeroDecimals(String.valueOf(femaleBinRanges.get(k+1)))));
-            }
-            femaleRangesInResults.add((">= " + trimTrailingZeroDecimals(String.valueOf(femaleBinRanges.get(femaleBinRanges.size()-1)))));
-
-            ArrayList<String> otherRangesInResults = new ArrayList<>();
-            otherRangesInResults.add(("< " + trimTrailingZeroDecimals(String.valueOf(otherBinRanges.get(0)))));
-            for (int k=0; k<otherBinRanges.size()-1;k++) {
-                otherRangesInResults.add((trimTrailingZeroDecimals(String.valueOf(otherBinRanges.get(k))) + " - " + trimTrailingZeroDecimals(String.valueOf(otherBinRanges.get(k+1)))));
-            }
-            otherRangesInResults.add((">= " + trimTrailingZeroDecimals(String.valueOf(otherBinRanges.get(otherBinRanges.size()-1)))));
-
-            for(AchillesResult ar: aa.getResults()){
-                String analysisStratumName=ar.getAnalysisStratumName();
-                String result_value = ar.getStratum4();
-
-                if(Long.valueOf(ar.getStratum3()) == MALE && maleRangesInResults.contains(result_value)){
-                    maleRangesInResults.remove(result_value);
-                }else if(Long.valueOf(ar.getStratum3()) == FEMALE && femaleRangesInResults.contains(result_value)){
-                    femaleRangesInResults.remove(result_value);
-                }else if(Long.valueOf(ar.getStratum3()) == OTHER && otherRangesInResults.contains(result_value)){
-                    otherRangesInResults.remove(result_value);
-                }
-
-                if (analysisStratumName == null || analysisStratumName.equals("")) {
-                    ar.setAnalysisStratumName(genderStratumNameMap.get(ar.getStratum3()));
-                }
-            }
-
-            for(String maleRemaining: maleRangesInResults){
-                String missingBinWidth = null;
-                if (maleBinWidth == (long)(maleBinWidth)) {
-                    missingBinWidth = String.format("%d",(long)maleBinWidth);
-                } else {
-                    missingBinWidth = String.format("%.2f", maleBinWidth);
-                }
-                missingBinWidth = trimTrailingZeroDecimals(missingBinWidth);
-                AchillesResult achillesResult = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, unitName, String.valueOf(MALE), maleRemaining, null, String.valueOf(maleBinWidth), null, 20L, 20L);
-                aa.addResultsItem(achillesResult);
-            }
-
-            for(String femaleRemaining: femaleRangesInResults){
-                String missingBinWidth = null;
-                if (femaleBinWidth == (long)femaleBinWidth) {
-                    missingBinWidth = String.format("%d",(long)femaleBinWidth);
-                } else {
-                    missingBinWidth = String.format("%.2f", femaleBinWidth);
-                }
-                missingBinWidth = trimTrailingZeroDecimals(missingBinWidth);
-                AchillesResult ar = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, unitName, String.valueOf(FEMALE), femaleRemaining, null, String.valueOf(femaleBinWidth), null, 20L, 20L);
-                aa.addResultsItem(ar);
-            }
-
-            for(String otherRemaining: otherRangesInResults){
-                String missingBinWidth = null;
-                if (otherBinWidth == (long)otherBinWidth) {
-                    missingBinWidth = String.format("%d",(long)otherBinWidth);
-                } else {
-                    missingBinWidth = String.format("%.2f", otherBinWidth);
-                }
-                missingBinWidth = trimTrailingZeroDecimals(missingBinWidth);
-                AchillesResult ar = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, unitName, String.valueOf(OTHER), otherRemaining, null, String.valueOf(otherBinWidth), null, 20L, 20L);
-                aa.addResultsItem(ar);
-            }
-        } else {
-
-            List<AchillesResult> maleResults = new ArrayList<>();
-            List<AchillesResult> femaleResults = new ArrayList<>();
-            List<AchillesResult> otherResults = new ArrayList<>();
-
-            for(AchillesResult ar: aa.getResults()){
-                String analysisStratumName=ar.getAnalysisStratumName();
-                if(Long.valueOf(ar.getStratum3()) == MALE ){
-                    maleResults.add(ar);
-                }else if(Long.valueOf(ar.getStratum3()) == FEMALE ){
-                    femaleResults.add(ar);
-                }else if(Long.valueOf(ar.getStratum3()) == OTHER ){
-                    otherResults.add(ar);
-                }
-                if (analysisStratumName == null || analysisStratumName.equals("")) {
-                    ar.setAnalysisStratumName(genderStratumNameMap.get(ar.getStratum2()));
-                }
-            }
-
-            if (("numeric").equals(type)) {
-                if (maleResults.size() == 0) {
-                    AchillesResult achillesResult = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, "No Unit", String.valueOf(MALE), "0", null, "0", null, 20L, 20L);
-                    aa.addResultsItem(achillesResult);
-                }
-                if (femaleResults.size() == 0) {
-                    AchillesResult ar = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, unitName, String.valueOf(FEMALE), "0", null, "0", null, 20L, 20L);
-                    aa.addResultsItem(ar);
-                }
-                if (otherResults.size() == 0) {
-                    AchillesResult ar = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, unitName, String.valueOf(OTHER), "0", null, "0", null, 20L, 20L);
-                    aa.addResultsItem(ar);
-                }
-            } else if(("text").equals(type)) {
-                if (maleResults.size() == 0) {
-                    AchillesResult achillesResult = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, "No Unit", String.valueOf(MALE), "Null", null, "0", null, 20L, 20L);
-                    aa.addResultsItem(achillesResult);
-                }
-                if (femaleResults.size() == 0) {
-                    AchillesResult ar = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, unitName, String.valueOf(FEMALE), "Null", null, "0", null, 20L, 20L);
-                    aa.addResultsItem(ar);
-                }
-                if (otherResults.size() == 0) {
-                    AchillesResult ar = achillesMapper.makeCopyAchillesResult(MEASUREMENT_GENDER_ANALYSIS_ID, conceptId, unitName, String.valueOf(OTHER), "Null", null, "0", null, 20L, 20L);
-                    aa.addResultsItem(ar);
-                }
-            }
-
-        }
-
-    }
-
-    public static HashMap<String,List<AchillesResult>> seperateUnitResults(Analysis aa){
-        List<String> distinctUnits = new ArrayList<>();
-
-        for(AchillesResult ar:aa.getResults()){
-            if(!distinctUnits.contains(ar.getStratum2()) && !Strings.isNullOrEmpty(ar.getStratum2())){
-                distinctUnits.add(ar.getStratum2());
-            }
-        }
-
-        Multimap<String, AchillesResult> resultsWithUnits = Multimaps
-                .index(aa.getResults(), AchillesResult::getStratum2);
-
-        HashMap<String,List<AchillesResult>> seperatedResults = new HashMap<>();
-
-        for(String key:resultsWithUnits.keySet()){
-            seperatedResults.put(key,new ArrayList<>(resultsWithUnits.get(key)));
-        }
-        return seperatedResults;
-    }
-
-
-    public static HashMap<String,List<AchillesResultDist>> seperateDistResultsByUnit(List<AchillesResultDist> results) {
-        Multimap<String, AchillesResultDist> distResultsWithUnits = Multimaps
-                .index(results, AchillesResultDist::getStratum2);
-        HashMap<String,List<AchillesResultDist>> seperatedResults = new HashMap<>();
-
-        for(String key:distResultsWithUnits.keySet()){
-            seperatedResults.put(key,new ArrayList<>(distResultsWithUnits.get(key)));
-        }
-
-        return seperatedResults;
-    }
-
-    public String trimTrailingZeroDecimals(String s) {
-        String trimmedValue = null;
-        if (s != null) {
-            trimmedValue = s.indexOf(".") < 0 ? s : s.replaceAll("0*$", "").replaceAll("\\.$", "");
-        }
-        return trimmedValue;
-    }
-
-    public float normalizeBinWidth(Float bMin, Float bMax) {
-        Float binWidth = (float)0;
-
-        if (((bMax - bMin)/11) >= 5) {
-            binWidth = (float) (Math.round(((bMax-bMin)/11)/5)*5);
-        } else if (((bMax-bMin)/11) >= 0.1 && ((bMax-bMin)/11) <= 1) {
-            binWidth = (float) (Math.ceil((float)(((bMax-bMin)/11))/0.1)*0.1);
-        } else if (((bMax-bMin)/11) >= 1 && ((bMax-bMin)/11) <= 2) {
-            binWidth = (float) (Math.round((float)(((bMax-bMin)/11))/2)*2);
-        } else if (((bMax-bMin)/11) >= 2 && ((bMax-bMin)/11) <= 3) {
-            binWidth = (float) (Math.round((float)(((bMax-bMin)/11))/3)*3);
-        } else if (((bMax-bMin)/11) >= 3 && ((bMax-bMin)/11) <= 5) {
-            binWidth = (float) (Math.round((float)(((bMax-bMin)/11))/5)*5);
-        } else {
-            binWidth = (float) (Math.round(((bMax-bMin)/11))*10.0/10.0);
-        }
-
-        return binWidth;
     }
 
     public List<QuestionConcept> mapAnalysesToQuestions(List<Analysis> analyses, List<QuestionConcept> questions) {
